backend/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.6
channels==4.0.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
pandas==1.5.3